refactor: move helper function into TopicService
