+++ src/test/java/de/caritas/cob/userservice/api/facade/CreateUserFacadeTest.java
@@ -30,9 +30,9 @@ import de.caritas.cob.userservice.api.exception.httpresponses.CustomValidationHt
 import de.caritas.cob.userservice.api.exception.httpresponses.InternalServerErrorException;
 import de.caritas.cob.userservice.api.facade.rollback.RollbackFacade;
 import de.caritas.cob.userservice.api.helper.AgencyVerifier;
-import de.caritas.cob.userservice.api.helper.RegistrationStatisticsHelper;
 import de.caritas.cob.userservice.api.helper.UserVerifier;
 import de.caritas.cob.userservice.api.manager.consultingtype.ConsultingTypeManager;
+import de.caritas.cob.userservice.api.service.consultingtype.TopicService;
 import de.caritas.cob.userservice.api.service.statistics.StatisticsService;
 import de.caritas.cob.userservice.api.service.user.UserService;
 import de.caritas.cob.userservice.consultingtypeservice.generated.web.model.ExtendedConsultingTypeResponseDTO;
@@ -56,8 +56,7 @@ public class CreateUserFacadeTest {
   @Mock private CreateNewConsultingTypeFacade createNewConsultingTypeFacade;
   @Mock private UserVerifier userVerifier;
   @Mock private StatisticsService statisticsService;
-
-  @Mock private RegistrationStatisticsHelper registrationStatisticsHelper;
+  @Mock private TopicService topicService;
 
   @Test
   public void
